requirements.txt
Flask
Flask_Caching
Flask_Cors
Flask_JWT_Extended
Flask_Mail
flask_paginate
Flask_Session
lxml
numpy
opencv_python
Pillow
PyYAML
redis
Requests
SQLAlchemy
torch
torchvision
Werkzeug
